requirements.txt
cryptography.fernet
Fernet
